LayoutTests/platform/win-7sp0/media/W3C/video/canPlayType/canPlayType_codecs_order_3-expected.txt

video.canPlayType() - codecs parameter order

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "probably" is "probably"

TEST COMPLETE
spec reference
